Dynamically load Emoji Mart
