+++ components/Feed.tsx
@@ -84,7 +84,7 @@ function Item({ item }: { item: FeedItem }): React.ReactElement {
 
       {
         object?.__typename === 'Comment' && item.activities.length === 1 && (
-          <div className="mt-4 border-l-2 border-primary bg-gray-100 text-md italic">
+          <div className="mt-4 border-l-2 border-primary bg-gray-100 text-md">
             <CommentEditor
               resourceId={object.resourceId}
               initialValue={JSON.parse(object.message)}
